requirements.txt
anthropic>=0.40.0
yfinance>=0.2.40
feedparser>=6.0.10
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
fredapi>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pykrx>=1.0.45
